requirements.txt
google-generativeai==0.4.1
python-dotenv==1.0.1
python-telegram-bot